atools: Added new script feedbackcount.py
The feedbackcount.py script will run similar to findecho.py
but instead identifying feedback and count peaks in those
The distance between the feedback peaks will be the roundtrip delay.
Currently the algorithm is blind and only looking at peak values and
filtering out to short and too long diffs.
In the future is could do a band pass filter for a feedback frequency
and count that instead.

Signed-off-by: JohanBlome <johan@gunhildcarling.net>