atools: Sharper filtering and histogram analysis
A few tweaks to limit potential noise polution.
1) the signal is hp (fs/40) and lp (fs/4)
2) the bandpass filter is higher q
3) the peak diffs are calculated at intervals and not just adjacent
4) the diffs are put ina histogram where the largest bin is selected
   and the length is the average of the edges.

Still things can go wrong. Most often as a result of to few feedback
echos, for which a warning is printed.
However, it may be that there are high enery periodic noise in the
background or some other reflective effect causing confusion in the
algorithm. Use verbose argument to see what is going on in the signal.

Signed-off-by: JohanBlome <johan@gunhildcarling.net>